refactor: Move layout code from Base to LayoutObserver, clean up
